django_apps/drf_swagger_docs_versioning/requirements.txt
Django==5.1
djangorestframework==3.15.2
drf-spectacular==0.27.2

